Switching to Maven and PaperAPI
