--- .idea/artifacts/Feudalism_jar.xml
@@ -1,8 +1,0 @@
-<component name="ArtifactManager">
-  <artifact type="jar" name="Feudalism:jar">
-    <output-path>$PROJECT_DIR$/</output-path>
-    <root id="archive" name="Feudalism.jar">
-      <element id="module-output" name="Feudalism" />
-    </root>
-  </artifact>
-</component>
