+++ .idea/libraries/Maven__com_destroystokyo_paper_paper_api_1_16_3_R0_1_SNAPSHOT.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: com.destroystokyo.paper:paper-api:1.16.3-R0.1-SNAPSHOT">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/com/destroystokyo/paper/paper-api/1.16.3-R0.1-SNAPSHOT/paper-api-1.16.3-R0.1-20201011.021843-33.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/com/destroystokyo/paper/paper-api/1.16.3-R0.1-SNAPSHOT/paper-api-1.16.3-R0.1-20201011.021843-33-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/com/destroystokyo/paper/paper-api/1.16.3-R0.1-SNAPSHOT/paper-api-1.16.3-R0.1-20201011.021843-33-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
