+++ .idea/libraries/Maven__com_github_LlmDl_Towny_0_96_2_19.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: com.github.LlmDl:Towny:0.96.2.19">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/com/github/LlmDl/Towny/0.96.2.19/Towny-0.96.2.19.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/com/github/LlmDl/Towny/0.96.2.19/Towny-0.96.2.19-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/com/github/LlmDl/Towny/0.96.2.19/Towny-0.96.2.19-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
