--- README.md
@@ -1,4 +1,0 @@
-## Feudalism
-A Stoneskies Plugin
-
-Adding complex war to towny
